# materialize-sql: use `encrow.Shape` for encoding