@williamhbaker 2023-11-10
One addendum: I'm swapping in `github.com/klauspost/compress/gzip` in place of the standard gzip package. This improves the encoding benchmarks for about 40% with compression enabled:

_New optimized code, with compression enabled using `github.com/klauspost/compress/gzip`:_
```
go test -v ./materialize-sql/ -bench=. -benchmem -benchtime=5s -count=3
BenchmarkEncodingObjects
BenchmarkEncodingObjects-10         2302           2528709 ns/op
BenchmarkEncodingObjects-10         2474           2435775 ns/op
BenchmarkEncodingObjects-10         2461           2417468 ns/op
BenchmarkEncodingArrays
BenchmarkEncodingArrays-10          2269           2667515 ns/op
BenchmarkEncodingArrays-10          2280           2653695 ns/op
BenchmarkEncodingArrays-10          2258           2660670 ns/op
```