@williamhbaker 2023-11-10
Another addendum: I had removed the `countingWriteCloser` proxy struct thinking it was a nice simplification, but it was actually necessary to output files at the target file size of _compressed_ bytes, so I added that back in.